cpu: Move all of the instructions into the cpu module
The more I thought about it, the less sense it made to keep all of the
instruction implementations outside of the CPU impl; logically it makes
sense, the first parameter to all instructions is a mutable reference to
the CPU, and the instructions all call internal CPU instructions.

This also allows me to make a bunch of public CPU functions non-public.