ines: only support mapper 0
In order to keep the scope more defined, I'm only going to support
mapper 0 for now. That will cover most of the games that I want to play,
and I can always refactor to add support for more mappers later anyway.
But for now I don't want to worry about it.