cpu: fixed my PPUADDR bug
For some reason, when I was incrementing PPUADDR on _both_ a PPUDATA
read and a write, the output would be broken. If I commented out the
increment during a PPUDATA read, then it would work.

This was happening because there was a read followed by a write because
the CPU was _always_ retrieving data from memory, assuming the read
wasn't destructive. So, when writing to PPUDATA, there would be a read
form $2007 (which incremented PPUADDR) and then the write to $2007 would
be executed (which incremented PPUADDR again).

This has fixed the rendering of Ice Climber's title screen, and it has
fixed _some_ of Super Mario Bros title screen.