core: implement rudimentary save states
I'm not 100% happy with hand rolling, but I couldn't make any of the
existing crates work how I wanted it to. `bincode' was the closest, but
had all kinds of borrowing issues for my code.

Anyway, so far it works. I've played a few days worth of Zelda and Super
Mario Bros without any issues.