core: refactor read/write methods away from Result type
In the beginning, using the Result type as a return type for my
read/write methods seemed to make sense, but the fact is that in the
hardware, if you read/write to/from memory that you're not meant to, the
hardware doesn't fail or kill the NES console or anything; it still
returns something. So bubbling up errors to cause a panic just takes
away from the experience, when the game would probably just continue to
run fine anyway.