docs: clarify that Sunsoft 5B games work, but without extra sound
Gimmick! works, but the extra sound channels that it utilises aren't
working. However its use of these extra channels is so minimal that you
barely notice. This is the only Sunsoft 5B game that uses the extra
sound channels, so as long as it's documented, I'm happy to list 5B
carts as supported.