apu: dmc: implemented the DMC channel
This is a mostly cycle accurate implementation of the DMC channel. The
way IRQs are triggered required some extra work to separate
frame-generated IRQs from DMC-generated IRQs.

For tricky games that heavily rely on the DMC IRQ for timing, this
probably isn't accurate enough, but for basic DMC use as an audio
channel, it works well enough.