mapper2: put PRG-RAM back into UxROM
A homebrew NES game (Alwa's Adventure) requires PRG-RAM to be present in
UxROM.

I don't see any mention of PRG-RAM being mapped for any mapper 2 PCBs,
but it's easy to put it back in.

Also, we're using all 8 bits of the register to swap the first PRG-ROM
bank, since we can, and homebrew games can use it.